+++ crypto/src/main/java/org/web3j/crypto/TransactionEncoder.java
@@ -17,8 +17,8 @@ import org.web3j.utils.Numeric;
  * <a href="http://gavwood.com/paper.pdf">yellow paper</a>.
  */
 public class TransactionEncoder {
-	
-	private static final int CHAIN_ID_INC = 35;
+
+    private static final int CHAIN_ID_INC = 35;
     private static final int LOWER_REAL_V = 27;
 
     public static byte[] signMessage(RawTransaction rawTransaction, Credentials credentials) {
@@ -30,7 +30,7 @@ public class TransactionEncoder {
     }
 
     public static byte[] signMessage(
-            RawTransaction rawTransaction, byte chainId, Credentials credentials) {
+            RawTransaction rawTransaction, long chainId, Credentials credentials) {
         byte[] encodedTransaction = encode(rawTransaction, chainId);
         Sign.SignatureData signatureData = Sign.signMessage(
                 encodedTransaction, credentials.getEcKeyPair());
@@ -40,14 +40,14 @@ public class TransactionEncoder {
     }
 
     public static Sign.SignatureData createEip155SignatureData(
-            Sign.SignatureData signatureData, byte chainId) {
+            Sign.SignatureData signatureData, long chainId) {
 //        byte v = (byte) (signatureData.getV() + (chainId << 1) + 8);
         BigInteger v =  Numeric.toBigInt(signatureData.getV());
         v = v.subtract(BigInteger.valueOf(LOWER_REAL_V));
         v = v.add(BigInteger.valueOf(chainId * 2));
         v = v.add(BigInteger.valueOf(CHAIN_ID_INC));
 
-        
+
 
         return new Sign.SignatureData(
                 v.toByteArray(), signatureData.getR(), signatureData.getS());
@@ -57,9 +57,9 @@ public class TransactionEncoder {
         return encode(rawTransaction, null);
     }
 
-    public static byte[] encode(RawTransaction rawTransaction, byte chainId) {
+    public static byte[] encode(RawTransaction rawTransaction, long chainId) {
         Sign.SignatureData signatureData = new Sign.SignatureData(
-        		longToBytes(chainId), new byte[] {}, new byte[] {});
+                longToBytes(chainId), new byte[] {}, new byte[] {});
         return encode(rawTransaction, signatureData);
     }
 
@@ -68,7 +68,7 @@ public class TransactionEncoder {
         RlpList rlpList = new RlpList(values);
         return RlpEncoder.encode(rlpList);
     }
-    
+
     private static byte[] longToBytes(long x) {
         ByteBuffer buffer = ByteBuffer.allocate(Long.BYTES);
         buffer.putLong(x);
